Refactor token sorting for external use
