+++ packages/ui-components/src/lib/wallet/index.ts
@@ -1,2 +1,3 @@
 export * from './login';
 export * from './signer';
+export * from './tokens';
